"Fixed" --events parameter for CLI
previously:
    khal agenda --events=N #didn't have any effect

now:
It does. To be fair: It doesn't limit the amount of events, but the
number of lines shown (i.e. day headings are included in count).

For me, this is a more desirable behaviour to define a maximum length of
output.

When both, --days and --events are specified, the more restrictive
option wins. Particularly, if no --days is specified, --days=2 is set.
Increasing default number of days might be usable:
    khal agenda --days=10 --events=N # could be useful